refactor(ddd): ddd 아키텍처로 변경
