+++ app/domain/entities/user.py
@@ -1,6 +1,6 @@
-from sqlalchemy import String, Integer, DateTime, func
+from sqlalchemy import String
 from sqlalchemy.orm import Mapped, mapped_column
-from app.db.session import Base
+from app.infrastructure.db.session import Base
 
 class User(Base):
     id: Mapped[int] = mapped_column(primary_key=True, index=True)
